Show round info card on the active round page
Extracts the wizard's confirmation card into a shared RoundInfoCard component
and reuses it in the active round page, replacing the separate host banner and
instructions callout.
